// frontend/src/api/models/MachineView.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type MachineView = {
    readonly id?: number;
    serNumM: string;
    readonly modelm_title?: string;
    readonly modele_title?: string;
    serNumE: string;
    readonly modelt_title?: string;
    serNumT: string;
    readonly modelda_title?: string;
    serNumDA: string;
    readonly modelsa_title?: string;
    serNumSA: string;
    NumSupContract: string;
    dateShipmentFactory: string;
    consignee: string;
    addressOperation: string;
    equipment: string;
    userClient: number;
    userService: number;
};